# src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mithra Direct App",
    "description": "API documentation for the subscription services",
    "contact": {
      "name": "Swamy Kunta",
      "url": "https://github.com/s713278/ECommerceApp/tree/master",
      "email": "swamy.kunta@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "/"
    },
    "version": "v1"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local Env"
    },
    {
      "url": "http://172.105.33.153:8080/api",
      "description": "Dev Env"
    }
  ],
  "paths": {
...
    "/v1/users/{userId}": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "summary": "User Information",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "put": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Update User Information",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User Update Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "delete": {
        "tags": [
          "5. User Management"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      },
      "patch": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Update Delivery Address",
        "description": "Updates the delivery address for a specific user. Only valid keys (address1, address2, city, state, zipCode, country) are accepted.",
        "operationId": "addNameAndAddress",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "A map containing valid delivery address fields. Valid keys: address1, address2, city, state, zipCode, country.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameAndAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Address updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input keys or validation error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "example": {
                    "error": "Invalid keys found: [invalidKey]"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/{userId}/subs/": {
      "post": {
        "tags": [
          "3. Subscription Management"
        ],
        "summary": "Create Subscription",
        "operationId": "createSubscription",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/{userId}/mobile": {
      "patch": {
        "tags": [
          "5. User Management"
        ],
        "operationId": "updateMobileAddress",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMobileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/users/{userId}/del_instructions": {
      "patch": {
        "tags": [
          "5. User Management"
        ],
        "operationId": "updateDeliveryInstructions",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/{userId}/vendor/{vendorId}": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Fetch Subscriptions By Vendor and User",
        "operationId": "fetchSubsByUserAndVendor",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
    "/v1/users/{userId}/subs": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "summary": "Fetch Subscriptions By User ID",
        "operationId": "fetchAllSubsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
    "/v1/users/admin": {
      "get": {
        "tags": [
          "5. User Management"
        ],
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "userId"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "E-Commerce Application": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponseObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Customer registered successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "description": "HTTP Status Code",
            "format": "int32",
            "example": 201
          },
          "data": {
            "type": "object",
            "description": "Data"
          }
        }
      },
...
      "CreateSubscriptionRequest": {
        "required": [
          "frequency",
          "price_id",
          "quantity",
          "sku_id"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "DAILY",
              "ALTERNATE_DAY",
              "WEEKLY",
              "CUSTOM"
            ]
          },
          "sku_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "price_id": {
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "delivery_date": {
            "type": "string",
            "format": "date"
          },
          "start_date": {
            "type": "string",
            "format": "date"
          },
          "end_date": {
            "type": "string",
            "format": "date"
          },
          "custom_days": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "Data": {
        "type": "object",
        "properties": {
          "subscription_id": {
            "type": "integer",
            "format": "int64"
          },
          "next_delivery_date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Data"
          }
        }
      },
...
      "NameAndAddressRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "object",
            "example":{
                "address1": "123 Main Street",
                "address2": "Apt 4B",
                "city": "Hyderabad",
                "state": "Telangana",
                "zipCode": "500001",
                "country": "India"
            }
          }
        }
      },
...
      "UpdateMobileRequest": {
        "required": [
          "new_mobile_number",
          "old_mobile_number"
        ],
        "type": "object",
        "properties": {
          "old_mobile_number": {
            "pattern": "^\\d{10}$",
            "type": "integer",
            "description": "Mobile number",
            "format": "int64",
            "example": 9876543210
          },
          "new_mobile_number": {
            "pattern": "^\\d{10}$",
            "type": "integer",
            "description": "Mobile number",
            "format": "int64",
            "example": 9090912345
          }
        },
        "description": "Update Mobile Request"
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User Update Request"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "lastPage": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "E-Commerce Application": {
        "type": "http",
        "in": "header",
        "scheme": "bearer"
      }
    }
  }
}